repackage source code and updated network tutorial
