+++ setup.py
@@ -1,0 +1,12 @@
+from setuptools import setup
+
+setup(
+    authors = "Russell Littman, Michael Cheng", email="rlittman16@g.ucla.edu",
+    author_email="rlittman16@g.ucla.edu, mcheng7777@g.ucla.edu",
+    name='SCING',
+    version="1.0.0",
+    description="Single Cell INtegrative Gene regulatory network inference",
+    py_modules=['SCING'],
+    package_dir={'':'src',
+                 },
+)
