+++ src/scing/__init__.py
@@ -1,0 +1,5 @@
+from . import (
+    supercellHelpers as supercells,
+    buildGRNHelpers as build,
+    MergeNetworksHelpers as merge
+)
